fix to work with struct

+++ matlab/low/2d/fleids2heid.m
@@ -3,4 +3,4 @@ function heid = fleids2heid(fid, leid) %#codegen
 % HEID = FLEIDS2HEID(FID, LEID)
 % See also HEID2FID, HEID2LEID
 
-heid = fid*4+leid-1;
+heid = fid*4+int32(leid)-1;
